Work toward a new, common, scalable manifest model.
Instead of having HTML5-specific and Gears-specific manifests that we load,
which contain every resource needed in offline mode, these new manifests are
custom and can be treated individually.

There's now a master manifests list that point to all the other manifests we
care about for an offline session. Each manifest is versioned and is in a
format similar to the Google Gears manifest format.

This will allow us to fetch the master list and then determine on a
per-manifest basis whether we need to load their URLs. Each sub-manifest is
dedicated to a review request and everything related to it, and is versioned
separately from the master manifest list.

The big benefit is that we won't necessarily have to fetch every single URL
when a user goes into offline mode, and one small change in a review request
won't trigger a full reload of all URLs, which is the main problem with a
single versioned manifest. Having these separate manifests in conjunction
with custom manifest loaders will also let us properly add progress
information and to have only one format we have to support.