Add a concept of URL aliases, and use the backend to copy the original URL for each alias.
We need this because we key off media files with the MEDIA_SERIAL, and while
this works well in most cases, we have problems with our CSS files which do not
reference by MEDIA_SERIAL.

With this, we can now alias all media files to include one with the serial
and one without. This is safe because we only access this cache when offline,
so we won't screw with online usage after upgrades by serving stale data.