Separate out the UI and model for diff comments.
RB.DiffComment now handles all data storage operations and uses events to
communicate text changes, saves, deletes, and more. Eventing is handled by
jQuery.event.

The DiffCommentBlock now stores one of these when there's a draft comment and
updates its UI based on the state of that comment.

$.fn.commentDlg now takes a comment list and a draft comment as separate
values. It operates only on the draft comment model and has no knowledge
of the UI side.

This introduces a fix for a bug where we weren't always updating the state of
the Save button in the comment dialog. It also has better, more consistent
updating for the color of the draft comments